requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
flexmatcher==1.0.4
idna==2.9
isodate==0.6.0
isort==4.3.21
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.3
mccabe==0.6.1
more-itertools==8.2.0
munkres==1.1.2
nltk==3.4.5
numpy==1.18.1
Owlready2==0.23
pandas==0.25.3
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
pywikibot==3.0.20200111
PyYAML==5.3
rdflib==4.2.2
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
simpy==3.0.11
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
urllib3==1.25.8
wrapt==1.11.2
